detekt fixes
