+++ config/detekt/detekt-baseline.xml
@@ -10,17 +10,14 @@
     <ID>EmptyFunctionBlock:listenerConversions.kt${ }</ID>
     <ID>LargeClass:Purchases.kt$Purchases$Purchases</ID>
     <ID>LongMethod:Purchases.kt$Purchases$restorePurchases</ID>
-    <ID>LongParameterList:Backend.kt$Backend$( purchaseToken: String, appUserID: String, isRestore: Boolean, observerMode: Boolean, subscriberAttributes: Map&lt;String, SubscriberAttribute&gt;, productInfo: ProductInfo, onSuccess: PostReceiptDataSuccessCallback, onError: PostReceiptDataErrorCallback )</ID>
+    <ID>LongParameterList:Backend.kt$Backend$( purchaseToken: String, appUserID: String, isRestore: Boolean, observerMode: Boolean, subscriberAttributes: Map&lt;String, Map&lt;String, Any?&gt;&gt;, productInfo: ProductInfo, onSuccess: PostReceiptDataSuccessCallback, onError: PostReceiptDataErrorCallback )</ID>
     <ID>LongParameterList:EntitlementInfo.kt$EntitlementInfo$( val identifier: String, val isActive: Boolean, val willRenew: Boolean, val periodType: PeriodType, val latestPurchaseDate: Date, val originalPurchaseDate: Date, val expirationDate: Date?, val store: Store, val productIdentifier: String, val isSandbox: Boolean, val unsubscribeDetectedAt: Date?, val billingIssueDetectedAt: Date? )</ID>
-    <ID>LongParameterList:PurchaserInfo.kt$PurchaserInfo$( val entitlements: EntitlementInfos, val purchasedNonSubscriptionSkus: Set&lt;String&gt;, val allExpirationDatesByProduct: Map&lt;String, Date?&gt;, val allPurchaseDatesByProduct: Map&lt;String, Date?&gt;, @Deprecated("Use getExpirationDateForEntitlement instead") val allExpirationDatesByEntitlement: Map&lt;String, Date?&gt;, @Deprecated("Use getPurchaseDateForEntitlement instead") val allPurchaseDatesByEntitlement: Map&lt;String, Date?&gt;, val requestDate: Date, internal val jsonObject: JSONObject, internal val schemaVersion: Int, val firstSeen: Date, val originalAppUserId: String )</ID>
-    <ID>LongParameterList:Purchases.kt$Purchases$( private val application: Application, backingFieldAppUserID: String?, private val backend: Backend, private val billingWrapper: BillingWrapper, private val deviceCache: DeviceCache, private val executorService: ExecutorService, private val identityManager: IdentityManager, private val subscriberAttributesManager: SubscriberAttributesManager, appConfig: AppConfig )</ID>
-    <ID>LongParameterList:Purchases.kt$Purchases$( purchase: PurchaseWrapper, skuDetails: SkuDetails?, allowSharingPlayStoreAccount: Boolean, consumeAllTransactions: Boolean, appUserID: String, onSuccess: ((PurchaseWrapper, PurchaserInfo) -&gt; Unit)? = null, onError: ((PurchaseWrapper, PurchasesError) -&gt; Unit)? = null )</ID>
-    <ID>LongParameterList:Purchases.kt$Purchases$( purchases: List&lt;PurchaseWrapper&gt;, allowSharingPlayStoreAccount: Boolean, consumeAllTransactions: Boolean, appUserID: String, onSuccess: ((PurchaseWrapper, PurchaserInfo) -&gt; Unit)? = null, onError: ((PurchaseWrapper, PurchasesError) -&gt; Unit)? = null )</ID>
-    <ID>MagicNumber:Purchases.kt$Purchases.AttributionNetwork.BRANCH$3</ID>
-    <ID>MagicNumber:Purchases.kt$Purchases.AttributionNetwork.FACEBOOK$5</ID>
-    <ID>MagicNumber:Purchases.kt$Purchases.AttributionNetwork.TENJIN$4</ID>
+    <ID>MagicNumber:build.gradle.kts$14</ID>
+    <ID>MagicNumber:build.gradle.kts$28</ID>
+    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendBadRequest$7226</ID>
     <ID>MagicNumber:errors.kt$BackendErrorCode.BackendCannotTransferPurchase$7102</ID>
     <ID>MagicNumber:errors.kt$BackendErrorCode.BackendEmptyAppUserId$7220</ID>
+    <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInternalServerError$7110</ID>
     <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAPIKey$7225</ID>
     <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAppStoreSharedSecret$7104</ID>
     <ID>MagicNumber:errors.kt$BackendErrorCode.BackendInvalidAppleSubscriptionKey$7234</ID>
@@ -40,14 +37,11 @@
     <ID>MagicNumber:utils.kt$1000000.0</ID>
     <ID>ReturnCount:BillingWrapper.kt$BillingWrapper$internal fun getPurchaseType(purchaseToken: String): PurchaseType</ID>
     <ID>TooGenericExceptionCaught:AdvertisingIdClient.kt$AdvertisingIdClient$e: Exception</ID>
-    <ID>TooGenericExceptionCaught:utils.kt$e: RuntimeException</ID>
     <ID>TooGenericExceptionThrown:HTTPClient.kt$HTTPClient$throw RuntimeException(e)</ID>
     <ID>TooManyFunctions:Backend.kt$Backend$Backend</ID>
     <ID>TooManyFunctions:BillingWrapper.kt$BillingWrapper$BillingWrapper</ID>
     <ID>TooManyFunctions:DeviceCache.kt$DeviceCache$DeviceCache</ID>
-    <ID>TooManyFunctions:Factories.kt$com.revenuecat.purchases.Factories.kt</ID>
     <ID>TooManyFunctions:Purchases.kt$Purchases$Purchases</ID>
     <ID>TooManyFunctions:listenerConversions.kt$com.revenuecat.purchases.listenerConversions.kt</ID>
-    <ID>TooManyFunctions:utils.kt$com.revenuecat.purchases.utils.kt</ID>
   </Whitelist>
 </SmellBaseline>
